refactor: rename DataSourceSwitcher to SourceSwitcher
