+++ packages/plugins/ENS/src/constants.ts
@@ -8,7 +8,7 @@ export const ENS_PluginID = PluginID.ENS
 
 export const ENS_CONTRACT_ADDRESS = '0x57f1887a8BF19b14fC0dF6Fd9B2acc9Af147eA85'
 
-export const SupportedProvider = [
+export const SUPPORTED_SOURCE_TYPES = [
     SourceType.OpenSea,
     SourceType.NFTScan,
     SourceType.Gem,
