+++ packages/shared/src/UI/components/SourceSwitcher/index.tsx
@@ -1,0 +1,67 @@
+import { makeStyles, useStylesExtends } from '@masknet/theme'
+import { Stack, Typography } from '@mui/material'
+import { SourceType, resolveSourceTypeName } from '@masknet/web3-shared-base'
+import { Box } from '@mui/system'
+import { useSharedI18N } from '@masknet/shared'
+import { FootnoteMenu, FootnoteMenuOption } from '../FootnoteMenu'
+import { SourceProviderIcon } from '../SourceProviderIcon'
+
+const useStyles = makeStyles()((theme) => {
+    return {
+        source: {
+            justifyContent: 'space-between',
+        },
+        sourceNote: {
+            color: theme.palette.maskColor.secondaryDark,
+            fontSize: 14,
+            fontWeight: 700,
+        },
+        sourceMenu: {
+            fontSize: 14,
+            fontWeight: 'bolder',
+        },
+        sourceName: {
+            fontSize: 14,
+            fontWeight: 700,
+            color: theme.palette.mode === 'dark' ? '' : theme.palette.maskColor.publicMain,
+        },
+    }
+})
+
+export interface SourceSwitcherProps extends withClasses<'source' | 'sourceNote'> {
+    sourceType?: SourceType
+    sourceTypes?: SourceType[]
+    onSourceTypeChange?: (option: FootnoteMenuOption) => void
+}
+
+export function SourceSwitcher(props: SourceSwitcherProps) {
+    const { sourceType, sourceTypes = [], onSourceTypeChange } = props
+    const t = useSharedI18N()
+    const classes = useStylesExtends(useStyles(), props)
+
+    return (
+        <Box className={classes.source}>
+            <Stack
+                className={classes.sourceMenu}
+                display="inline-flex"
+                flexDirection="row"
+                alignItems="center"
+                gap={0.5}>
+                <Typography className={classes.sourceNote}>{t.plugin_trader_data_source()}</Typography>
+                <FootnoteMenu
+                    options={sourceTypes.map((x) => ({
+                        name: (
+                            <Stack display="inline-flex" flexDirection="row" alignItems="center" gap={0.5}>
+                                <SourceProviderIcon provider={x} size={20} />
+                                <Typography className={classes.sourceName}>{resolveSourceTypeName(x)}</Typography>
+                            </Stack>
+                        ),
+                        value: x,
+                    }))}
+                    selectedIndex={typeof sourceType !== 'undefined' ? sourceTypes.indexOf(sourceType) : -1}
+                    onChange={onSourceTypeChange}
+                />
+            </Stack>
+        </Box>
+    )
+}
