+++ packages/shared/src/UI/components/index.ts
@@ -30,7 +30,7 @@ export * from './PluginCardFrame/index.js'
 export * from './Markdown/index.js'
 export * from './Icon/index.js'
 export * from './FungibleTokenInput/index.js'
-export * from './DataSourceSwitcher/index.js'
+export * from './SourceSwitcher/index.js'
 export * from './FootnoteMenu/index.js'
 export * from './MaskPluginWrapper/index.js'
 export * from './SourceProviderIcon/index.js'
